+++ internal/doc_definition_test.go
@@ -21,6 +21,7 @@
 package internal
 
 import (
+	"fmt"
 	"sort"
 	"strings"
 )
@@ -56,4 +57,9 @@ func (d DocDefinitions) Sort() {
 		}
 		return a < b
 	})
+}
+
+func getDocsLinkForField(docName string, def DocDefinition) string {
+	anchor := fmt.Sprintf("%s-%s", strings.ToLower(strings.ReplaceAll(def.Path, ".", "")), strings.ToLower(def.Type))
+	return fmt.Sprintf("https://arangodb.github.io/kube-arangodb/docs/api/%s#%s", docName, anchor)
 }