+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - (Feature) License Manager
 - (Improvement) Use Async mode for backup creation
 - (Feature) (ML) CRD
+- (Feature) Add validation rules for CRDs
 
 ## [1.2.35](https://github.com/arangodb/kube-arangodb/tree/1.2.35) (2023-11-06)
 - (Maintenance) Update go-driver to v1.6.0, update IsNotFound() checks